+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.109.0
+
+- Upgrade framer-motion to latest version.
+- DraggableEntity component to encapsulate logic of updating location with framer-motion drag.
+- Widget component now to use DraggableEntity.
+- No longer using `initialOffset` property in widgets; now setting initial position when they are created.
+- Moving the `location` property of widgets into the `data` sub-property from the top level widget data.
+
 ## 1.108.0
 
 - fixing sticky note to always display who added it, even after that user leaves the room.